+++ lib/Sort/quickSort.h
@@ -34,8 +34,8 @@ namespace AEPKSS::Sort
     };
 }
 
-static void sort(vector<size_t> &in, size_t left, size_t right);
-static size_t partition(vector<size_t> &in, size_t left, size_t right);
+static void sort(vector<size_t> &in, long long left, long long right);
+static size_t partition(vector<size_t> &in, long long left, long long right);
 
 static size_t partition_parallel(vector<size_t> &in, size_t concurrency);
 static AEPKSS::Sort::ParallelQuickSortReturn sort_parallel(vector<size_t> &in, size_t pivot);
